+++ wolfssl/src/wifi.c
@@ -106,7 +106,7 @@ static void handle_wifi_disconnect_result(struct net_mgmt_event_callback *cb) {
 }
 
 void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb,
-                             uint32_t mgmt_event, struct net_if *iface) {
+                             uint64_t mgmt_event, struct net_if *iface) {
     switch (mgmt_event) {
     case NET_EVENT_WIFI_SCAN_RESULT:
         handle_wifi_scan_result(cb);
